bisq-desktop: utilize the built-in Tor handling
Bisq comes with an embedded Tor binary (provided by a third party), but
we don't use it in this package because it's build for a FHS-abiding
Linux distro; Meaning, Tor won't execute because it tried to load
libraries from standard locations.

To address this problem, the Nix package launches an ephemeral Tor
instance for Bisq. The approach works, but it does mean having to manage the tor process, something which is already handled well by Bisq.

This change modifies the Bisq Jar archive such that it launches the Tor
binary from Nixpkgs, allowing Bisq to manage Tor as it does on other
Linux distros and operating systems.