sparrow: init at 1.4.3
This is a semi-working version of Sparrow.

The problem is with Java's module system. The Sparrow distribution
packages the app using jlink, which merges the necessary components of
the JRE and the app-specific module into a module (JIMAGE). Using the
app-specific module causes problems because the included Linux libraries
are not set up to run on Nix.

There's a tool called jimage which can be used to inspect and extract
JIMAGEs.